fix: refactoring Category

+++ API_Project_PM.Core/Services/Categories/CategoryService.cs
@@ -1,4 +1,6 @@
-﻿using API_Project_PM.Core.Database;
+﻿using API_Project_PM.Core.CustomException;
+using API_Project_PM.Core.CustomExceptions;
+using API_Project_PM.Core.Database;
 using API_Project_PM.Core.Models;
 using Microsoft.EntityFrameworkCore;
 
@@ -26,15 +28,19 @@ namespace API_Project_PM.Core.Services.Categories
 
         public async Task<Category> CreateAsync(Category item)
         {
-                _db.Categories.Add(item);
-                await _db.SaveChangesAsync();
-                return item;
+            bool duplicatName = await _db.Categories.AnyAsync(c => c.Name == item.Name);
+            if (duplicatName) throw new ConflictException($"Categorie met naam: {item.Name} bestaat al!");
+
+            _db.Categories.Add(item);
+            await _db.SaveChangesAsync();
+            return item;
         }
         public async Task<bool> UpdateAsync(Category item)
         {
-            var result = await _db.Categories.FindAsync(item.Id);
+            Category? result = await _db.Categories.FindAsync(item.Id);
             if (result is null) return false;
 
+            if (await _db.Categories.AnyAsync(c => c.Name == item.Name && c.Id != item.Id)) throw new ConflictException($"Categorie met naam: {item.Name} bestaat al!");
 
             _db.Entry(result).CurrentValues.SetValues(item);
 
@@ -45,8 +51,12 @@ namespace API_Project_PM.Core.Services.Categories
 
         public async Task<bool> DeleteAsync(int id)
         {
-            var result = await _db.Categories.FindAsync(id);
+            Category? result = await _db.Categories.FindAsync(id);
             if (result is null) return false;
+
+            bool categoryHasActiveParts = await _db.Parts.IgnoreQueryFilters().AnyAsync(p => p.CategoryId == result.Id);
+
+            if (categoryHasActiveParts) throw new CannotDeleteException($"Categorie {result.Name}", "Parts");
 
             _db.Categories.Remove(result);
